+++ pkg/util/time.go
@@ -5,10 +5,13 @@ import (
 	"math/rand"
 	"net/http"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/pkg/errors"
 	"github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/promql"
+	"github.com/prometheus/prometheus/promql/parser"
 	"github.com/weaveworks/common/httpgrpc"
 )
 
@@ -33,6 +36,10 @@ func FormatTimeMillis(ms int64) string {
 // FormatTimeModel returns a human readable version of the input time.
 func FormatTimeModel(t model.Time) string {
 	return TimeFromMillis(int64(t)).String()
+}
+
+func FormatMillisToSeconds(ms int64) string {
+	return strconv.FormatFloat(float64(ms)/float64(1000), 'f', -1, 64)
 }
 
 // ParseTime parses the string into an int64, milliseconds since epoch.
@@ -97,4 +104,35 @@ func NewDisableableTicker(interval time.Duration) (func(), <-chan time.Time) {
 
 	tick := time.NewTicker(interval)
 	return func() { tick.Stop() }, tick.C
+}
+
+// FindMinMaxTime returns the time in milliseconds of the earliest and latest point in time the statement will try to process.
+// This takes into account offsets, @ modifiers, and range selectors.
+// If the expression does not select series, then FindMinMaxTime returns (0, 0).
+func FindMinMaxTime(r *http.Request, expr parser.Expr, lookbackDelta time.Duration, now time.Time) (int64, int64) {
+	isQuery := strings.HasSuffix(r.URL.Path, "/query")
+
+	var startTime, endTime int64
+	if isQuery {
+		if t, err := ParseTimeParam(r, "time", now.UnixMilli()); err == nil {
+			startTime = t
+			endTime = t
+		}
+	} else {
+		if st, err := ParseTime(r.FormValue("start")); err == nil {
+			if et, err := ParseTime(r.FormValue("end")); err == nil {
+				startTime = st
+				endTime = et
+			}
+		}
+	}
+
+	es := &parser.EvalStmt{
+		Expr:          expr,
+		Start:         TimeFromMillis(startTime),
+		End:           TimeFromMillis(endTime),
+		LookbackDelta: lookbackDelta,
+	}
+
+	return promql.FindMinMaxTime(es)
 }